backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

PyMySQL==1.1.1

marshmallow==3.21.1
flask-marshmallow==1.2.1

PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-dotenv==1.0.1
flask-cors==4.0.1
Flask-Limiter==3.7.0
pytest==8.1.1

gunicorn==22.0.0

